spark/requirements.txt
requests
beautifulsoup4
confluent-kafka
psycopg2-binary
